generate 2 CSV files of 10M data, proper PostgreSQL schema
