+++ package.json
@@ -8,7 +8,7 @@
     "node": ">=6.13.0"
   },
   "scripts": {
-    "seed": "node server/seed.js",
+    "seed": "node database/seed.js",
     "test": "jest",
     "build": "webpack",
     "start": "nodemon server/index.js",
@@ -25,9 +25,11 @@
     "mysql": "^2.18.1",
     "nodemon": "^2.0.4",
     "path": "^0.12.7",
+    "pg": "^8.3.3",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "readline": "^1.3.0",
     "styled-components": "^5.1.1",
     "webpack": "^4.44.1"
   },
